api/requirements.txt
astroid==2.4.2
flask-cors
attrs==19.3.0
autopep8==1.5.3
bcrypt
cffi==1.14.0
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-json-schema==0.0.5
Flask-JWT-Extended==3.24.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pyrsistent==0.16.0
python-dotenv==0.13.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
debugpy
mysql-connector-python
